Output from sensor now displayed on page. Issue was due to the method signature being slightly incorrect.
CURRENT ISSUES
Display problems with layouts.

+++ ExpressiveTouch/SensorData.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 struct SensorData {
-    let x:Int, y:Int, z:Int
+    let x:Double, y:Double, z:Double
 }
 
 enum SensorDataType:Int {
